chore: clean up unused code rename sections to be more accurate
