--- frontend/src/pages/Register.jsx
@@ -1,147 +1,0 @@
-import React, { useState, useEffect } from "react";
-import { useSelector, useDispatch } from "react-redux";
-import { useNavigate } from "react-router-dom";
-import { toast } from "react-toastify";
-import { FaUser } from "react-icons/fa";
-import { register, reset } from "../features/auth/authSlice";
-import Spinner from "../components/Spinner";
-
-function Register() {
-  const [formData, setFormData] = useState({
-    name: "",
-    email: "",
-    password: "",
-    password2: "",
-  });
-
-  const { name, email, password, password2 } = formData;
-
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
-
-  const { user, isLoading, isError, isSuccess, message } = useSelector(
-    (state) => state.auth
-  );
-
-  useEffect(() => {
-    if (isError) {
-      toast.error(message, {
-        position: toast.POSITION.TOP_RIGHT,
-        className: "alert alert-error",
-      });
-    }
-
-    if (isSuccess || user) {
-      navigate("/oldDash");
-    }
-
-    dispatch(reset());
-  }, [user, isError, isSuccess, message, navigate, dispatch]);
-
-  const onChange = (e) => {
-    setFormData((prevState) => ({
-      ...prevState,
-      [e.target.name]: e.target.value,
-    }));
-  };
-
-  const onSubmit = (e) => {
-    e.preventDefault();
-
-    if (password !== password2) {
-      toast.error("Passwords do not match!");
-    } else {
-      const userData = {
-        name,
-        email,
-        password,
-      };
-
-      dispatch(register(userData));
-    }
-  };
-
-  if (isLoading) {
-    return <Spinner />;
-  }
-
-  return (
-    <>
-      <section className="py-12">
-        <h1 className="text-4xl text-center">
-          <FaUser className="inline pb-2" /> Signup
-        </h1>
-        <p className="text-center">Please create an account</p>
-      </section>
-
-      <section className="flex justify-center">
-        <form onSubmit={onSubmit}>
-          <div className="form-control">
-            <label className="label flex-col items-start px-0" htmlFor="name">
-              <p className="label-text px-4">Your Name </p>
-              <input
-                type="text"
-                className="input input-bordered justify-center w-full"
-                id="name"
-                name="name"
-                value={name}
-                placeholder="Enter your name"
-                onChange={onChange}
-              />
-            </label>
-            <label className="label flex-col items-start px-0" htmlFor="email">
-              <p className="label-text px-4">Your Email </p>
-              <input
-                type="email"
-                className="input input-bordered justify-center w-full"
-                id="email"
-                name="email"
-                value={email}
-                placeholder="Enter your email"
-                onChange={onChange}
-              />
-            </label>
-            <label
-              className="label flex-col items-start px-0"
-              htmlFor="password"
-            >
-              <p className="label-text px-4">Password </p>
-              <input
-                type="password"
-                className="input input-bordered justify-center w-full"
-                id="password"
-                name="password"
-                value={password}
-                placeholder="Enter password"
-                onChange={onChange}
-              />
-            </label>
-            <label
-              className="label flex-col items-start px-0"
-              htmlFor="password2"
-            >
-              <p className="label-text px-4">Confirm Password </p>
-              <input
-                type="password"
-                className="input input-bordered justify-center w-full"
-                id="password2"
-                name="password2"
-                value={password2}
-                placeholder="Confirm password"
-                onChange={onChange}
-              />
-            </label>
-            <button
-              type="submit"
-              className="btn btn-accent btn-wide mx-auto my-4"
-            >
-              Signup
-            </button>
-          </div>
-        </form>
-      </section>
-    </>
-  );
-}
-
-export default Register;
